+++ java/src/test/java/org/rocksdb/CompactionOptionsFIFOTest.java
@@ -18,9 +18,27 @@ public class CompactionOptionsFIFOTest {
   @Test
   public void maxTableFilesSize() {
     final long size = 500 * 1024 * 1026;
-    try(final CompactionOptionsFIFO opt = new CompactionOptionsFIFO()) {
+    try (final CompactionOptionsFIFO opt = new CompactionOptionsFIFO()) {
       opt.setMaxTableFilesSize(size);
       assertThat(opt.maxTableFilesSize()).isEqualTo(size);
+    }
+  }
+
+  @Test
+  public void ttl() {
+    final long ttl = 7 * 24 * 60 * 60; // 7 days
+    try (final CompactionOptionsFIFO opt = new CompactionOptionsFIFO()) {
+      opt.setTtl(ttl);
+      assertThat(opt.ttl()).isEqualTo(ttl);
+    }
+  }
+
+  @Test
+  public void allowCompaction() {
+    final boolean allowCompaction = true;
+    try (final CompactionOptionsFIFO opt = new CompactionOptionsFIFO()) {
+      opt.setAllowCompaction(allowCompaction);
+      assertThat(opt.allowCompaction()).isEqualTo(allowCompaction);
     }
   }
 }