+++ table/block_based_table_reader.h
@@ -114,7 +114,7 @@ class BlockBasedTable : public TableReader {
                                 bool skip_filters = false,
                                 bool for_compaction = false) override;
 
-  InternalIterator* NewRangeTombstoneIterator(
+  FragmentedRangeTombstoneIterator* NewRangeTombstoneIterator(
       const ReadOptions& read_options) override;
 
   // @param skip_filters Disables loading/accessing the filter block
@@ -322,9 +322,9 @@ class BlockBasedTable : public TableReader {
       CachableEntry<Block>* block, BlockContents* raw_block_contents,
       CompressionType raw_block_comp_type, uint32_t format_version,
       const Slice& compression_dict, SequenceNumber seq_no,
-      size_t read_amp_bytes_per_bit, bool is_index = false,
-      Cache::Priority pri = Cache::Priority::LOW,
-      GetContext* get_context = nullptr, MemoryAllocator* allocator = nullptr);
+      size_t read_amp_bytes_per_bit, MemoryAllocator* memory_allocator,
+      bool is_index = false, Cache::Priority pri = Cache::Priority::LOW,
+      GetContext* get_context = nullptr);
 
   // Calls (*handle_result)(arg, ...) repeatedly, starting with the entry found
   // after a call to Seek(key), until handle_result returns false.